% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{hError}
\alias{hError}
\title{Hierarchical Error-Based Clustering}
\usage{
hError(data, errors, siglevel = 0.05, verbose = T)
}
\arguments{
\item{data}{Array of measurements.}

\item{errors}{Array of errors. Must have same length as \emph{data}.}

\item{siglevel}{(Optional) Significance level for Chi-Squared test. Defaults to 0.05.}

\item{verbose}{(Optional) Boolean to trigger printing to screen. Defaults to \emph{TRUE}.}
}
\description{
Generalization of Ward's hierarchical clustering algorithm for error-based clustering, developed by Kumar and Patel (2007).
}
\examples{
data <- c(rnorm(50, 3, 2), rnorm(50, 0, 5))
error <- c(rep(2, 50), rep(5, 50))
test <- hError(data, error)

test$metadata$cluster
test$metadata$classification
}
\seealso{
\code{\link[clustErrors]{kError}}
}
