% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{kError}
\alias{kError}
\title{Error-Based Clustering for known number of clusters}
\usage{
kError(data, errors, k, verbose = T)
}
\arguments{
\item{data}{Array of measurements.}

\item{errors}{Array of errors. Must have same length as \emph{data}.}

\item{k}{Number of clusters.}

\item{verbose}{(Optional) Boolean to trigger printing to screen. Defaults to \emph{TRUE}.}
}
\description{
Generalization of k-means clustering algorithm for error-based clustering, developed by Kumar and Patel (2007).
}
\examples{
# Comparison with k-means when measurements all have same error
data <- c(rnorm(15, 3, 2), rnorm(15, 0, 5))
error <- rep(2, 30)
test0 <- kmeans(data, 2)
test1 <- kError(data, error, 2)

# results are the same, since errors are the same
test0$centers
test1$metadata$cluster

test0$cluster
test1$metadata$classification

# Comparison when measurements have varying errors
error <- runif(30, 0, 5)
test2 <- kError(data, error, 2)

# results are not the same, since errors are not the same
test0$centers
test2$metadata$cluster

test0$cluster
test2$metadata$classification
}
\seealso{
\code{\link[clustErrors]{hError}}
}
